+++ app/core/SDKConnect/SDKConnectConstants.ts
