+++ app/core/SDKConnect/handlers/handleSendMessage.ts
@@ -5,7 +5,7 @@ import Logger from '../../../util/Logger';
 import Device from '../../../util/device';
 import { Minimizer } from '../../NativeModules';
 import { Connection } from '../Connection';
-import { METHODS_TO_DELAY, RPC_METHODS } from '../SDKConnect.constants';
+import { METHODS_TO_DELAY, RPC_METHODS } from '../SDKConnectConstants';
 import DevLogger from '../utils/DevLogger';
 import { wait } from '../utils/wait.util';
 import handleBatchRpcResponse from './handleBatchRpcResponse';
